Move CategoricalEncoder to encoder::categorical
